fix: Fix CodeQL warning

+++ src/_data2d.hpp
@@ -10,13 +10,13 @@ public:
   Data2d(Pos2i const &start, Pos2i const &end, T def)
       : fStart(start), //
         fEnd(end) {
-    fStorage.resize((end.fX - start.fX + 1) * (end.fZ - start.fZ + 1), def);
+    fStorage.resize((size_t)(end.fX - start.fX + 1) * (end.fZ - start.fZ + 1), def);
   }
 
   Data2d(Pos2i const &start, u32 width, u32 height, T def)
       : fStart(start), //
         fEnd(start.fX + (int)width - 1, start.fZ + (int)height - 1) {
-    fStorage.resize(width * height, def);
+    fStorage.resize((size_t)width * height, def);
   }
 
   typename std::vector<T>::reference operator[](Pos2i const &p) {
